Fix: Applied reviewers changes
