+++ libs/material/src/lib/delivery/delivery-view/delivery-view.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { Observable } from 'rxjs';
 import { DeliveryQuery, DeliveryService } from '../+state';
-import { Movie, MovieQuery, MovieService, Stakeholder } from '@blockframes/movie';
-
+import { Movie, MovieQuery } from '@blockframes/movie';
+import { Material } from '../../material/+state';
 
 @Component({
   selector: 'delivery-view',
@@ -11,26 +11,27 @@ import { Movie, MovieQuery, MovieService, Stakeholder } from '@blockframes/movie
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DeliveryViewComponent implements OnInit {
-  public movie$: Observable<Movie>;
-  public stakeholders$: Observable<Stakeholder[]>;
+  public movie: Movie;
   public materials$: Observable<Object>;
   public progressionValue$: Observable<number>;
 
   constructor(
-    private movieService: MovieService,
     private movieQuery: MovieQuery,
-    private query: DeliveryQuery,
-    ) {}
+    private service: DeliveryService,
+    private deliveryQuery: DeliveryQuery
+  ) {}
 
   ngOnInit() {
-    this.movie$ = this.movieQuery.selectActive();
-    this.stakeholders$ = this.movieService.activeMovieStakeholders;
-    this.materials$ = this.query.materialsByActiveDelivery$;
-    this.progressionValue$ = this.query.deliveryProgression$;
+    this.movie = this.movieQuery.getActive();
+    this.materials$ = this.deliveryQuery.materialsByActiveDelivery$;
+    this.progressionValue$ = this.deliveryQuery.deliveryProgression$;
+  }
+
+  public deliveredToggle(material: Material, movieId: string) {
+    this.service.deliveredToggle(material, movieId).catch(err => console.log(err));
   }
 
   public randomNumberPicker(scale: number) {
     return Math.floor(Math.random() * scale) + 1;
   }
-
 }
