+++ libs/material/src/lib/delivery/movie-material-item/movie-material-item.component.html
@@ -1,49 +1,30 @@
-<mat-card
-[ngClass]="{'materials-disabled': !!disableMaterial}"
-class="materials-block mat-elevation-z6"
-fxLayout="row"
-fxLayoutGap="20px">
+<mat-card class="materials-block mat-elevation-z6" fxLayout="row" fxLayoutGap="20px">
   <mat-checkbox
     fxLayoutAlign="center center"
     color="primary"
     [(ngModel)]="material.delivered"
-    [disabled]="!!disableMaterial"
-    (click)="!disableMaterial && deliveredToggle(material)"
+    (click)="deliveredToggle(material)"
   >
   </mat-checkbox>
   <div class="material" fxLayout="column" fxLayoutAlign="center start">
     <span>
       <h4 mat-line>{{ material.value }}</h4>
     </span>
-    <mat-expansion-panel [ngClass]="{'expansion-panel-disabled': !!disableMaterial}" [expanded]="isOpen" class="mat-elevation-z0">
+    <mat-expansion-panel [expanded]="isOpen" class="mat-elevation-z0">
       <p mat-line class="description-text">{{ material.description }}</p>
     </mat-expansion-panel>
-    <span fxLayout="row" fxLayoutAlign="start center">
-      <button [ngClass]="{'toggle-disabled': !!disableMaterial}" class="toggle mat-elevation-z0" mat-raised-button (click)="!disableMaterial && panelToggle()">
-        {{ panelButtonLabel }}
-      </button>
-      <span class="divider">|</span>
-      <button [ngClass]="{'disable-disabled': !!disableMaterial}" class="disable mat-elevation-z0" mat-raised-button (click)="disableToggle()">
-        {{ disableButtonLabel }}
-      </button>
-    </span>
+    <button class="toggle mat-elevation-z0" mat-raised-button (click)="panelToggle()">
+      {{ panelButtonLabel }}
+    </button>
   </div>
   <div class="informations" fxLayout="row" fxLayoutAlign="space-around none">
-    <span *ngIf="!disableMaterial" class="material-step">
-      <h5>Step :</h5>
-      <p class="material-step-text">Step 5</p>
+    <span class="material-step">
+      <h5>Ordered by :</h5>
+      <p class="material-step-text">{{ stakeholder }}</p>
     </span>
     <span class="state">
       <h5>State :</h5>
       <mat-icon class="icons" svgIcon="{{ stateIcon }}"></mat-icon>
-    </span>
-    <span *ngIf="!disableMaterial" class="time">
-      <h5>Delivery time :</h5>
-      <p class="time-text">{{ randomNumberPicker(30) }} days</p>
-    </span>
-    <span class="payment">
-      <h5>Payment :</h5>
-      <mat-icon class="icons" svgIcon="{{ paymentIcon }}"></mat-icon>
     </span>
   </div>
 </mat-card>
