+++ libs/material/src/lib/delivery/movie-material-item/movie-material-item.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, ChangeDetectionStrategy, Input } from '@angular/core';
 import { Material } from '../../material/+state';
 import { DeliveryService } from '../+state';
-import { MovieQuery } from '@blockframes/movie';
+import { MovieQuery, Stakeholder } from '@blockframes/movie';
+import { OrganizationQuery } from '@blockframes/organization';
 
 @Component({
   selector: 'delivery-movie-material-item',
@@ -11,51 +12,42 @@ import { MovieQuery } from '@blockframes/movie';
 })
 export class MovieMaterialItemComponent implements OnInit {
   @Input() material: Material;
+  @Input() stakeholders: Stakeholder[];
 
   // Visual bullshit for WoW effect
-  public stateIcons = ['dummy', 'accepted', 'available', 'delivered', 'pending', 'refused'];
-  public paymentIcons = ['dummy', 'payed', 'not_payed'];
+  public stateIcons = ['accepted', 'available', 'delivered', 'pending', 'refused'];
   public stateIcon: string;
-  public paymentIcon: string;
+  public stakeholder: string;
 
   public isOpen = true;
-  public disableMaterial = false;
-  public panelButtonLabel = 'LESS'
-  public disableButtonLabel = 'DISABLE'
+  public panelButtonLabel = 'LESS';
 
   constructor(
     private service: DeliveryService,
     private movieQuery: MovieQuery,
-    ) {}
+    private orgQuery: OrganizationQuery
+  ) {}
 
   ngOnInit() {
     this.stateIcon = this.stateIcons[this.randomNumberPicker(5)];
-    this.paymentIcon = this.paymentIcons[this.randomNumberPicker(2)];
+    this.stakeholder = this.orgQuery.getEntity(
+      this.stakeholders[this.randomNumberPicker(this.stakeholders.length)].orgId
+    ).name;
   }
 
   public deliveredToggle(material: Material) {
     this.service
       .deliveredToggle(material, this.movieQuery.getActiveId())
       .catch(err => console.log(err));
+    //TODO: Ask why we put .catch ? And why we put it without "try" before
   }
 
   public randomNumberPicker(scale: number) {
-    return Math.floor(Math.random() * scale) + 1;
+    return Math.floor(Math.random() * scale);
   }
 
   public panelToggle() {
     this.isOpen = !this.isOpen;
-    !!this.isOpen
-      ? this.panelButtonLabel = 'LESS'
-      : this.panelButtonLabel = 'MORE';
-  }
-
-  public disableToggle() {
-    this.disableMaterial = !this.disableMaterial;
-    this.isOpen = false;
-    this.panelButtonLabel = 'MORE'
-    !!this.disableMaterial
-      ? this.disableButtonLabel = 'ENABLE'
-      : this.disableButtonLabel = 'DISABLE';
+    !!this.isOpen ? (this.panelButtonLabel = 'LESS') : (this.panelButtonLabel = 'MORE');
   }
 }
