+++ styles/theme.scss
@@ -106,5 +106,5 @@ $blockframes-theme-warn: mat-palette($mat-bf-crimson);
 $blockframes-theme: mat-light-theme(
   $blockframes-theme-primary,
   $blockframes-theme-accent,
-  $blockframes-theme-warn
+  $blockframes-theme-warn,
 );
